+++ .github/workflows/release.yml
@@ -13,4 +13,3 @@ jobs:
         with:
           github-token: ${{ secrets.YDB_PLATFORM_BOT_TOKEN_REPO }}
           npm-token: ${{ secrets.NODE_AUTH_TOKEN }}
-          npm-dist-tag: rc
