+++ package.json
@@ -42,7 +42,7 @@
     "luxon": "^3.2.1",
     "reflect-metadata": "^0.1.13",
     "uuid": "^8.3.2",
-    "ydb-sdk-proto": "^1.2.1"
+    "ydb-sdk-proto": "^1.1.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.6.1",
