+++ src/scheme.ts
@@ -12,6 +12,7 @@ import {Logger} from './logging';
 import DiscoveryService, {Endpoint} from './discovery';
 import {retryable} from "./retries";
 import {ISslCredentials} from './ssl-credentials';
+import {OperationParamsSettings} from './table';
 
 import SchemeServiceAPI = Ydb.Scheme.V1.SchemeService;
 import ListDirectoryResult = Ydb.Scheme.ListDirectoryResult;
@@ -21,7 +22,6 @@ import IMakeDirectoryRequest = Ydb.Scheme.IMakeDirectoryRequest;
 import IPermissions = Ydb.Scheme.IPermissions;
 import {util} from "protobufjs";
 import EventEmitter = util.EventEmitter;
-import {OperationParamsSettings} from "./table/session";
 
 
 function preparePermissions(action?: IPermissions | null) {
@@ -70,7 +70,7 @@ interface ISchemeClientSettings {
     logger: Logger;
 }
 
-export class SchemeClient extends EventEmitter {
+export default class SchemeClient extends EventEmitter {
     private schemeServices: Map<Endpoint, SchemeService>;
 
     constructor(private settings: ISchemeClientSettings) {
