--- src/table/index.ts
@@ -1,4 +1,0 @@
-export * from './session-pool';
-export * from './session';
-export * from './table-client';
-export * from './table-session';
